Make specific directories and move files in Components
