+++ src/Components/Main/Header/Header.js
@@ -1,5 +1,5 @@
 import styled from "styled-components"
-import Color from "../../Constants/Color"
+import Color from "../../../Constants/Color"
 
 const HeaderWrapper = styled.div`
   background: ${Color.yellow};
